feat(nx-dev): refactor theme detection with useIsomorphicLayoutEffect
Replaced `useEffect` with `useIsomorphicLayoutEffect` for theme detection to support both client and server environments. Deprecated methods are used to support older Safari versions. Renamed some hooks for better consistency and clarity.
