+++ nx-dev/ui-primitives/src/index.ts
@@ -1,3 +1,4 @@
 export * from './lib/cx';
-export * from './lib/use-layout.effect';
-export * from './lib/use-media-query';
+export * from './lib/use-isomorphic-layout-effect';
+export * from './lib/use-layout-effect';
+export * from './lib/use-media-query-effect';
