+++ nx-dev/ui-primitives/src/lib/use-isomorphic-layout-effect.ts
@@ -1,0 +1,5 @@
+'use client';
+import { useEffect, useLayoutEffect } from 'react';
+
+export const useIsomorphicLayoutEffect =
+  typeof window !== 'undefined' ? useLayoutEffect : useEffect;
