+++ nx-dev/ui-primitives/src/lib/use-media-query-effect.ts
@@ -1,7 +1,7 @@
 'use client';
 import { useEffect, useState } from 'react';
 
-export function useMediaQuery(query: string): boolean {
+export function useMediaQueryEffect(query: string): boolean {
   const getMatches = (query: string): boolean => {
     // Prevents SSR issues
     if (typeof window !== 'undefined') {
